+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowDarkTitleBarDemo.axaml.cs
@@ -1,6 +1,3 @@
-using Avalonia.Controls;
-using Avalonia.Input;
-using Avalonia.Interactivity;
 using Ursa.Controls;
 
 namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
@@ -10,34 +7,5 @@ public partial class UrsaWindowDarkTitleBarDemo : UrsaWindow
     public UrsaWindowDarkTitleBarDemo()
     {
         InitializeComponent();
-        WindowRegionHelper.AttachRoundedCorners(this);
-    }
-
-    private void TitleBar_OnPointerPressed(object? sender, PointerPressedEventArgs e)
-    {
-        if (e.Source is Button)
-        {
-            return;
-        }
-
-        if (e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
-        {
-            BeginMoveDrag(e);
-        }
-    }
-
-    private void Minimize_OnClick(object? sender, RoutedEventArgs e)
-    {
-        WindowState = WindowState.Minimized;
-    }
-
-    private void MaximizeRestore_OnClick(object? sender, RoutedEventArgs e)
-    {
-        WindowState = WindowState == WindowState.Maximized ? WindowState.Normal : WindowState.Maximized;
-    }
-
-    private void Close_OnClick(object? sender, RoutedEventArgs e)
-    {
-        Close();
     }
 }